requirements.txt
pytest==9.0.2
allure-pytest==2.15.3
requests==2.32.5
jsonschema==4.26.0
pydantic==2.12.5
python-dotenv==1.2.1
selene==2.0.0rc9
selenium==4.41.0
Appium-Python-Client==5.2.6
ruff==0.15.2
mypy==1.19.1
types-requests==2.32.4.20260107
black==26.1.0

allure-python-commons==2.15.3
